+++ src/components/IdealImage/index.tsx
@@ -1,8 +1,9 @@
 import React from 'react';
 import Image from '@theme/IdealImage';
 import styles from './styles.module.css'
-export default function IdealImage({img, alt}) {
+
+export default function IdealImage({ img, alt }) {
   return (
-    <Image className={styles.idealImage} shouldAutoDownload={() => true}  img={img} alt={alt}/>
+    <Image className={styles.idealImage} shouldAutoDownload={() => true} img={img} alt={alt} />
   );
 }