+++ CHANGELOG.md
@@ -13,14 +13,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Added the "interoperability with external credentials" page;
 - Use-cases:
   - Added the "Age restricted content" page;
-  - Added plugin-ideal-image for image optimization
 
 ### Changed
 
 - Products/WorldCoin State Replication:
   - Updated the relayer version in the replication guide;
-- Use-cases:
-  - Used `@docusaurus/plugin-ideal-image` for image optimization
+- Used `@docusaurus/plugin-ideal-image` for image optimization
 
 ## [2.0.5] - 23-02-2024