+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Ecosystem/Freedom Tool:
   - Added "Overview" page;
   - Added "Smart contracts reference" page;
+- Whitepaper v3;
 
 ### Changed